refactor to remove logic out of __init__
- closes #3
